Refactor joker to use better packaging
