+++ src/main/java/javatro/core/jokers/addmult/HalfJoker.java
@@ -1,7 +1,8 @@
-package javatro.core.jokers;
+package javatro.core.jokers.addmult;
 
 import javatro.core.Card;
 import javatro.core.Score;
+import javatro.core.jokers.Joker;
 
 /** Represents a HalfJoker Joker. */
 public class HalfJoker extends Joker {
